fixed warning (mostly static cast)

+++ ICEngine/src/Core/Converter_private.h
@@ -37,10 +37,10 @@ inline SDL_Rect ICE_Convert_BoxToSDL(ICE_Box *box)
 	SDL_Rect sdl = { 0 };
 	if (box)
 	{
-		sdl.x = box->x; 
-		sdl.y = box->y; 
-		sdl.w = box->w; 
-		sdl.h = box->h;
+		sdl.x = (int)box->x; 
+		sdl.y = (int)box->y; 
+		sdl.w = (int)box->w; 
+		sdl.h = (int)box->h;
 	}
 		
 	return sdl;
